% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amerdigest-package.R
\docType{package}
\name{amerdigest-package}
\alias{amerdigest}
\alias{amerdigest-package}
\title{amerdigest: Mobility in the Americas Digest}
\description{
Generates report with charts for a given country using a prescribed template.
}
\author{
\strong{Maintainer}: Kenneth S. Reyes \email{kereyes@iom.int}

}
\keyword{internal}
