% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_plot.R
\name{digest_plot}
\alias{digest_plot}
\title{Generate a preset plot}
\usage{
digest_plot(key, iso = NULL, basesize = basesize, font = font, ...)
}
\arguments{
\item{key}{Plot key.}

\item{iso}{Country.}

\item{basesize}{Base text size. Default is 7.}

\item{font}{Font family. Default is Open Sans.}

\item{...}{Optional arguments.}
}
\value{
\code{ggplot2} object or a list of metadata.
}
\description{
If \code{iso=NULL}, function returns \code{key} plot's metadata.
}
