% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{irreg}
\alias{irreg}
\title{Irregular Entries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10438 rows and 4 columns.
}
\source{
IOM Lighthouse Project; various governments.
}
\usage{
irreg
}
\description{
Monthly count of irregular entries from selected American countries.
}
\details{
\describe{
\item{t}{month-year of count}
\item{geo}{reporting country (ISO-3)}
\item{nat}{nationality (ISO-3)}
\item{n}{count of individuals}
}

Compiled from published administrative data through the IOM
Lighthouse Project.
}
\keyword{datasets}
