% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_font_installed.R
\name{package_font_installed}
\alias{package_font_installed}
\title{Check if Open Sans font family is installed in the system}
\usage{
package_font_installed()
}
\value{
TRUE if Open Sans is installed in the system
}
\description{
Verify if you have Open Sans font installed in your system
}
\examples{
test <- systemfonts::system_fonts() |>
              dplyr::filter(  family == "Gill Sans Nova"  )
test
package_font_installed()
}
