% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_font_registered.R
\name{package_font_registered}
\alias{package_font_registered}
\title{Check if Open Sans font is registered through \code{extrafont}}
\usage{
package_font_registered()
}
\value{
TRUE if OpenSans, Arial and Gill Sans Nova are registered
}
\description{
Verify if Open Sans is registered through \code{extrafont}
}
\examples{
package_font_registered()
}
