% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl}
\alias{pl}
\alias{pl1}
\alias{pl_usd}
\alias{pl_pct}
\alias{pl_pct1}
\title{Format numbers with pretty labels}
\usage{
pl(num)

pl1(num)

pl_usd(num)

pl_pct(num)

pl_pct1(num)
}
\arguments{
\item{num}{Numeric vector or single number to format.}
}
\value{
Character vector of formatted numbers.
}
\description{
These wrapper functions apply the \code{prettylabel} function with preset options
for formatting numbers, currencies, and percentages in plots or tables.
}
\details{
Note: This assumes you have a function named \code{prettylabel} available in your environment.
If not, please provide or load it.
}
\examples{
# Example usage if prettylabel is defined
pl(c(1234, 56789))
}
