% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_label.R
\name{plot_label}
\alias{plot_label}
\title{Add Label to Top of ggplot Object}
\usage{
plot_label(plot, label, span = 2, h = 0.06, basesize = 7, font = "Open Sans")
}
\arguments{
\item{plot}{A ggplot object to be labeled.}

\item{label}{A string to display as the label.}

\item{span}{Integer (1, 2, or 3) controlling the width of the label box.}

\item{h}{Numeric height of the label box relative to the plot (default = 0.06).}

\item{basesize}{Numeric. Base font size in points. Default is 7.}

\item{font}{Character. Font family for all text elements. Default is "Gill Sans Nova".}
}
\value{
A ggdraw object (from the \code{cowplot} package) with the label applied.
}
\description{
Adds a styled label with a colored background to the top of a plot using \code{cowplot}.
Useful for annotating subplots with country names, indicators, or categories.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
plot_label(p, "Fuel Efficiency", span = 2)
}
