% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regflow.R
\name{plot_regflow}
\alias{plot_regflow}
\title{Plot Regular Entries and Exits}
\usage{
plot_regflow(
  iso,
  basesize = 7,
  font = "Open Sans",
  title = "Regular entries and exits"
)
}
\arguments{
\item{iso}{ISO3 country code.}

\item{basesize}{Base font size. Default is 7.}

\item{font}{Font family. Default is "Open Sans".}

\item{title}{Plot title. Default is "Regular entries and exits".}
}
\value{
A ggplot2 object.
}
\description{
Plot Regular Entries and Exits
}
\examples{
plot_regflow("PAN")
}
