% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regnet.R
\name{plot_regnet}
\alias{plot_regnet}
\title{Plot Net Regular Entries}
\usage{
plot_regnet(
  iso,
  basesize = 7,
  font = "Open Sans",
  title = "Net regular entries"
)
}
\arguments{
\item{iso}{ISO3 country code.}

\item{basesize}{Base font size. Default is 7.}

\item{font}{Font family. Default is "Open Sans".}

\item{title}{Plot title. Default is "Net regular entries".}
}
\value{
A ggplot2 object.
}
\description{
Plot Net Regular Entries
}
\examples{
plot_regnet("BRA", title = "Custom Net Entry Title")
}
