% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{regout}
\alias{regout}
\title{Regular Exits}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43431 rows and 4 columns.
}
\source{
IOM Lighthouse Project; various governments.
}
\usage{
regout
}
\description{
Monthly count of regular exits from selected American countries.
}
\details{
\describe{
\item{t}{month-year of count}
\item{geo}{reporting country (ISO-3)}
\item{nat}{nationality (ISO-3)}
\item{n}{count of individuals}
}

Compiled from published administrative data through the IOM
Lighthouse Project.
}
\keyword{datasets}
