% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_digest.R
\name{render_digest}
\alias{render_digest}
\title{Create a Quarto file based on defined template and for a specific country}
\usage{
render_digest(file_name = NULL, country_iso)
}
\arguments{
\item{file_name}{A character string specifying the name of the new
Quarto Markdown file.
If \code{NULL}, the function use \verb{"index}" as the default file name.}

\item{country_iso}{the country iso number for the report to generate}
}
\value{
This function does not return a value. It performs file operations to set up
the Quarto IOM story template in the current working directory.
}
\description{
This function creates a Quarto file based on defined template
by copying the necessary template files
into the \verb{_extensions} directory and generating
a new Quarto Markdown (\code{.qmd}) file.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if \code{file_name} is provided. If not, it will be index per default.
\item Verifies what \code{ext_name} is selected.
\item Creates the \verb{_extensions} directory if it does not exist.
\item Creates a subdirectory within \verb{_extensions} for the specified extension.
\item Copies the extension files from the package's internal \code{extdata}
directory to \verb{_extensions}.
\item Checks if the extension files were successfully copied.
\item Creates a new Quarto Markdown file based on the template.
\item Opens the new file in the editor.
}
}
\examples{
# render_digest(country_iso="ARG")
}
