% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amerdigest-package.R
\docType{package}
\name{amerdigest-package}
\alias{amerdigest}
\alias{amerdigest-package}
\title{amerdigest: Mobility in the Americas Digest}
\description{
Generates report with charts for a given country using a prescribed template.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iom.github.io/amerdigest/}
}

}
\author{
\strong{Maintainer}: Kenneth S. Reyes \email{kereyes@iom.int}

Authors:
\itemize{
  \item Freddy Nicolas Marquez \email{fmarquez@iom.int}
  \item Edouard Legoupil \email{elegoupil@iom.int}
}

Other contributors:
\itemize{
  \item IOM [copyright holder]
}

}
\keyword{internal}
