% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_data.R
\name{digest_data}
\alias{digest_data}
\title{Get data for plot}
\usage{
digest_data(key, iso = NULL, ...)
}
\arguments{
\item{key}{Get key.}

\item{iso}{Country.}

\item{...}{Optional arguments.}
}
\value{
List.
}
\description{
Get data for plot
}
